rename startegy to just 'membership'
