+++ src/strategies/membership/README.md
@@ -1,4 +1,4 @@
-# membership-based-voting
+# membership
 
 This strategy allows you to combine any arbitrary "membership" strategy with any "voting power" strategy. The idea is that you can only vote if you pass the membership strategy. This is super useful to enable things like quadratic voting or other non plutocratic systems that require sybil resistance. For example, Compound could say that voting power is your COMP balance, but only if you also possess a [UID](https://etherscan.io/address/0xba0439088dc1e75F58e0A7C107627942C15cbb41) or a PUNK (`erc721` strategy). Or you must pass BrightID verification (`brightid` strategy). Or make your own verification strategy based on on-chain behavior! The sky is the limit!
